+++ app/src/main/java/com/soquipment/android/SoquipmentApplication.kt
@@ -1,0 +1,14 @@
+package com.soquipment.android
+
+import android.app.Application
+import com.naver.maps.map.NaverMapSdk
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class SoquipmentApplication : Application() {
+    override fun onCreate() {
+        super.onCreate()
+        NaverMapSdk.getInstance(this).client =
+            NaverMapSdk.NaverCloudPlatformClient(BuildConfig.NAVER_CLIENT_ID)
+    }
+}
